PR#118
Moved AsyncSQLClient and SQLClient in the same sql.py file inside clients module.
