+++ application_sdk/activities/metadata_extraction/sql.py
@@ -7,7 +7,7 @@ from temporalio import activity
 from application_sdk import activity_pd
 from application_sdk.activities import ActivitiesInterface
 from application_sdk.activities.utils import get_workflow_id
-from application_sdk.clients.sql_client import SQLClient
+from application_sdk.clients.sql import SQLClient
 from application_sdk.handlers.sql import SQLHandler
 from application_sdk.inputs.json import JsonInput
 from application_sdk.inputs.statestore import StateStore
