+++ application_sdk/handlers/sql.py
@@ -7,7 +7,7 @@ import pandas as pd
 
 from application_sdk import activity_pd
 from application_sdk.application.fastapi.models import MetadataType
-from application_sdk.clients.sql_client import SQLClient
+from application_sdk.clients.sql import SQLClient
 from application_sdk.common.logger_adaptors import AtlanLoggerAdapter
 from application_sdk.handlers import HandlerInterface
 from application_sdk.workflows.sql.workflows.workflow import SQLWorkflow
