+++ examples/application_sql.py
@@ -35,7 +35,7 @@ from urllib.parse import quote_plus
 from application_sdk.activities.metadata_extraction.sql import (
     SQLMetadataExtractionActivities,
 )
-from application_sdk.clients.async_sql_client import AsyncSQLClient
+from application_sdk.clients.sql import AsyncSQLClient
 from application_sdk.clients.temporal_client import TemporalClient
 from application_sdk.common.logger_adaptors import AtlanLoggerAdapter
 from application_sdk.handlers.sql import SQLHandler
