+++ tests/unit/handlers/test_metadata.py
@@ -4,7 +4,7 @@ from unittest.mock import AsyncMock, MagicMock, call
 import pytest
 
 from application_sdk.app.rest.fastapi.models.workflow import MetadataType
-from application_sdk.clients.sql_client import SQLClient
+from application_sdk.clients.sql import SQLClient
 from application_sdk.handlers.sql import SQLHandler
 
 T = TypeVar("T")
